fixed rocBLAS test
The tests need to be run manually since a GPU is needed to run this test
and it can't be found from inside the sandbox. The test-binary+data is
added to the output when building with doCheck=true (it is false by
default), this leads to an additional 650MB, so it only makes sense for
testing. The test-binary is wrapped and can be called from the
result-directory created by nix-build. (Be aware the test take a long
time >6h for me).